+++ lib/uniris_lib.dart
@@ -2,7 +2,8 @@
 
 export 'src/api.dart';
 export 'src/crypto.dart';
-export 'src/key_pair.dart';
-export 'src/secret.dart';
+export 'src/model/key_pair.dart';
+export 'src/model/secret.dart';
 export 'src/transaction_builder.dart';
 export 'src/utils.dart';
+export 'src/address_util.dart';