+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+#### Version 0.0.4+4 (2021-05-04) 
+* Add api requests (fetchBalance, getTransactionContent, getTransactions)
+* Organize packages
+
 #### Version 0.0.3+3 (2021-04-21) 
 * Dartdoc
 * Change project structure